Couple of TODO notes on better API use
We are listing projects and images just
to verify if they are already present.

As the number of relevant projects is 1
and images 2 (control plane and worker),
we don't have to list and filter (for listing APIs
use limits/pagination/cursors). We can call
`view` endpoint with the resource name directly.

This commit just leaves TODO notes on this.